refactor: extract parser and index modules from email_imap_adapter.py to eliminate Qlty total-complexity smell
